PIM-6353: ES normalization - move pv in dedicated folder
