+++ src/Pim/Component/Catalog/Normalizer/Indexing/Product/CompletenessCollectionNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Doctrine\Common\Collections\Collection;
 use Pim\Component\Catalog\Model\CompletenessInterface;
