--- src/Pim/Component/Catalog/Normalizer/Indexing/Product/OptionNormalizer.php
@@ -1,38 +1,0 @@
-<?php
-
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
-
-use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Value\OptionValueInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-
-/**
- * Normalizer for an option product value
- *
- * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- */
-class OptionNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsNormalization($data, $format = null)
-    {
-        return $data instanceof OptionValueInterface && 'indexing' === $format;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getNormalizedData(ValueInterface $value)
-    {
-        $data = $value->getData();
-        if (null !== $data) {
-            return $data->getCode();
-        }
-
-        return null;
-    }
-}
