+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/BooleanNormalizer.php
@@ -1,19 +1,17 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Model\ValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Normalizer for a date product value
- *
- * @author    Anaël Chardan <anael.chardan@akeneo.com>
+ * @author    Damien Carcel (damien.carcel@akeneo.com)
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class DateNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
+class BooleanNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
@@ -21,7 +19,7 @@ class DateNormalizer extends AbstractProductValueNormalizer implements Normalize
     public function supportsNormalization($data, $format = null)
     {
         return $data instanceof ValueInterface &&
-            AttributeTypes::BACKEND_TYPE_DATE === $data->getAttribute()->getBackendType() &&
+            AttributeTypes::BACKEND_TYPE_BOOLEAN === $data->getAttribute()->getBackendType() &&
             'indexing' === $format;
     }
 
@@ -30,12 +28,6 @@ class DateNormalizer extends AbstractProductValueNormalizer implements Normalize
      */
     protected function getNormalizedData(ValueInterface $value)
     {
-        $date = $value->getData();
-
-        if ($date instanceof \DateTimeInterface) {
-            return $date->format('Y-m-d');
-        }
-
-        return null;
+        return $value->getData();
     }
 }
