+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/DummyNormalizer.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Value\OptionsValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Normalizer for a options product value
+ * Dummy normalizer for a product value
+ * TODO TIP-706: Workaround to normalize
  *
- * @author    Anaël Chardan <anael.chardan@akeneo.com>
+ * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class OptionsNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
+class DummyNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof OptionsValueInterface && 'indexing' === $format;
+        return $data instanceof ValueInterface && 'indexing' === $format;
     }
 
     /**
@@ -28,6 +28,6 @@ class OptionsNormalizer extends AbstractProductValueNormalizer implements Normal
      */
     protected function getNormalizedData(ValueInterface $value)
     {
-        return $value->getOptionCodes();
+        return null;
     }
 }
