+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/MetricNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Pim\Component\Catalog\Model\ValueInterface;
 use Pim\Component\Catalog\Value\MetricValueInterface;
