+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/NumberNormalizer.php
@@ -1,19 +1,21 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Model\ValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Normalizer for a text (simple text) product value
+ * Normalizer for a decimal or an integer product value
  *
  * @author    Samir Boulil <samir.boulil@akeneo.com>
+ * @author    Anaël Chardan <anael.chardan@akeneo.com>
+ * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class TextNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
+class NumberNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
@@ -21,7 +23,7 @@ class TextNormalizer extends AbstractProductValueNormalizer implements Normalize
     public function supportsNormalization($data, $format = null)
     {
         return $data instanceof ValueInterface &&
-            AttributeTypes::BACKEND_TYPE_TEXT === $data->getAttribute()->getBackendType() &&
+            AttributeTypes::BACKEND_TYPE_DECIMAL === $data->getAttribute()->getBackendType() &&
             'indexing' === $format;
     }
 
@@ -30,6 +32,12 @@ class TextNormalizer extends AbstractProductValueNormalizer implements Normalize
      */
     protected function getNormalizedData(ValueInterface $value)
     {
-        return $value->getData();
+        $number = $value->getData();
+
+        if (null !== $number) {
+            return (string) $number;
+        }
+
+        return null;
     }
 }
