+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/OptionNormalizer.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Pim\Component\Catalog\Model\ValueInterface;
+use Pim\Component\Catalog\Value\OptionValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Dummy normalizer for a product value
- * TODO TIP-706: Workaround to normalize
+ * Normalizer for an option product value
  *
  * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class DummyNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
+class OptionNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof ValueInterface && 'indexing' === $format;
+        return $data instanceof OptionValueInterface && 'indexing' === $format;
     }
 
     /**
@@ -28,6 +28,11 @@ class DummyNormalizer extends AbstractProductValueNormalizer implements Normaliz
      */
     protected function getNormalizedData(ValueInterface $value)
     {
+        $data = $value->getData();
+        if (null !== $data) {
+            return $data->getCode();
+        }
+
         return null;
     }
 }
