+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/OptionsNormalizer.php
@@ -1,26 +1,26 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
-use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Model\ValueInterface;
+use Pim\Component\Catalog\Value\OptionsValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * @author    Damien Carcel (damien.carcel@akeneo.com)
+ * Normalizer for a options product value
+ *
+ * @author    Anaël Chardan <anael.chardan@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class BooleanNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
+class OptionsNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof ValueInterface &&
-            AttributeTypes::BACKEND_TYPE_BOOLEAN === $data->getAttribute()->getBackendType() &&
-            'indexing' === $format;
+        return $data instanceof OptionsValueInterface && 'indexing' === $format;
     }
 
     /**
@@ -28,6 +28,6 @@ class BooleanNormalizer extends AbstractProductValueNormalizer implements Normal
      */
     protected function getNormalizedData(ValueInterface $value)
     {
-        return $value->getData();
+        return $value->getOptionCodes();
     }
 }
