+++ src/Pim/Component/Catalog/Normalizer/Indexing/ProductValue/TextNormalizer.php
@@ -1,21 +1,19 @@
 <?php
 
-namespace Pim\Component\Catalog\Normalizer\Indexing\Product;
+namespace Pim\Component\Catalog\Normalizer\Indexing\ProductValue;
 
 use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Model\ValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * Normalizer for a decimal or an integer product value
+ * Normalizer for a text (simple text) product value
  *
  * @author    Samir Boulil <samir.boulil@akeneo.com>
- * @author    Anaël Chardan <anael.chardan@akeneo.com>
- * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class NumberNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
+class TextNormalizer extends AbstractProductValueNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
@@ -23,7 +21,7 @@ class NumberNormalizer extends AbstractProductValueNormalizer implements Normali
     public function supportsNormalization($data, $format = null)
     {
         return $data instanceof ValueInterface &&
-            AttributeTypes::BACKEND_TYPE_DECIMAL === $data->getAttribute()->getBackendType() &&
+            AttributeTypes::BACKEND_TYPE_TEXT === $data->getAttribute()->getBackendType() &&
             'indexing' === $format;
     }
 
@@ -32,12 +30,6 @@ class NumberNormalizer extends AbstractProductValueNormalizer implements Normali
      */
     protected function getNormalizedData(ValueInterface $value)
     {
-        $number = $value->getData();
-
-        if (null !== $number) {
-            return (string) $number;
-        }
-
-        return null;
+        return $value->getData();
     }
 }
