+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/CompletenessCollectionNormalizerSpec.php
@@ -7,7 +7,7 @@ use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\ChannelInterface;
 use Pim\Component\Catalog\Model\CompletenessInterface;
 use Pim\Component\Catalog\Model\LocaleInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\Product\CompletenessCollectionNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\ProductValue\CompletenessCollectionNormalizer;
 use Prophecy\Argument;
 
 class CompletenessCollectionNormalizerSpec extends ObjectBehavior
