+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/MetricNormalizerSpec.php
@@ -6,7 +6,7 @@ use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\MetricInterface;
 use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\Product\MetricNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\ProductValue\MetricNormalizer;
 use Pim\Component\Catalog\Value\MetricValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
