+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/NumberNormalizerSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Component\Catalog\Normalizer\Indexing\Product;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\Product\NumberNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\ProductValue\NumberNormalizer;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 class NumberNormalizerSpec extends ObjectBehavior
