+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/OptionNormalizerSpec.php
@@ -6,7 +6,7 @@ use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\AttributeOptionInterface;
 use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\Product\OptionNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\ProductValue\OptionNormalizer;
 use Pim\Component\Catalog\Value\OptionValueInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
