+++ src/Pim/Component/Catalog/spec/Normalizer/Indexing/Product/PriceCollectionNormalizerSpec.php
@@ -6,7 +6,7 @@ use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\ProductPrice;
 use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\Product\PriceCollectionNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\ProductValue\PriceCollectionNormalizer;
 use Pim\Component\Catalog\Value\PriceCollectionValue;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
