+++ src/Pim/Component/ReferenceData/Normalizer/Indexing/Product/ReferenceDataNormalizer.php
@@ -3,7 +3,7 @@
 namespace Pim\Component\ReferenceData\Normalizer\Indexing\Product;
 
 use Pim\Component\Catalog\Model\ValueInterface;
-use Pim\Component\Catalog\Normalizer\Indexing\Product\AbstractProductValueNormalizer;
+use Pim\Component\Catalog\Normalizer\Indexing\ProductValue\AbstractProductValueNormalizer;
 use Pim\Component\ReferenceData\Value\ReferenceDataValue;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
